Bootstrap a working hiera config on a fresh install
After applying a few patches from Puppetlabs, I have a way to get hiera
bootstrapped on this Vagrant node.

It takes two runs to get into convergence, but it does eventually work.
The first run gets hiera installed and configured well enough to not
fail on lookup, and the second one gets ZFS on linux itself to do the
install

Now with this out of the way I can go back to actually developing the
module.
 * https://github.com/puppetlabs/puppetlabs-stdlib/pull/59
 * https://github.com/puppetlabs/hiera-puppet/pull/23